+++ vscode-ext/glyph-middleware.cjs
@@ -1560,7 +1560,7 @@ ${parsed.dynamicLine}`
   _createSourceMap(preservePrivacy = false) {
     return {
       privacy: preservePrivacy ? this.sourceMap?.privacy || [] : [],
-      version: "1.34.1",
+      version: "1.34.2",
       level: this.level,
       provider: this.provider,
       profile: this.providerProfile,